[Perl] Use new class structure to remove duplication
This commit makes use of the new structure introduced in the
previous commit ("Refactor module structure and split primitives")
primarily to simplify the checks in Avro::BinaryEncoder, which now
validate all data using the individual schema's 'is_data_valid'.

Before this change, there were data validation rules in both the
encoder and in the underlying schemas, and they didn't always line
up (for example, the encoder checked integers using a regular
expression and a range check, while the schema checked using a
'pack'-based roundtrip), opening the door for cases where the
encoder would accept data that the schema would reject.